git_is_touched: Don't use git status --porcelain (#11)
* git_is_touched: Don't use git status --porcelain

That's the single most expensive git call you can make - it has to
scan the entire repo for *everything*.

This changes the time for e.g. a Qt clone from 500ms to 100ms.

* git_is_touched: Just call is_staged and is_dirty

Note that the order is semi-important, because if staged already
returns true this is done in like 3ms.